requirements.txt
Django>3.0
psycopg2
asn1crypto
Babel
certifi
cffi
chardet
configparser
cryptography
django-haystack
django-phonenumber-field
enum34
et-xmlfile
future
idna
jdcal
lxml
oauthlib
openpyxl
phonenumbers
protobuf
pycparser
PyJWT
pyOpenSSL
PySocks
python-dateutil
pytz
requests
requests-oauthlib
six
urllib3
Whoosh
xlrd